+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/implementation/util/BuilderHelper.java
@@ -37,7 +37,7 @@ import com.azure.storage.common.implementation.credentials.CredentialValidator;
 import com.azure.storage.common.policy.MetadataValidationPolicy;
 import com.azure.storage.common.policy.RequestRetryOptions;
 import com.azure.storage.common.policy.ResponseValidationPolicyBuilder;
-import com.azure.storage.common.policy.ScrubEtagPolicy;
+import com.azure.storage.common.policy.StorageEtagPolicy;
 import com.azure.storage.common.policy.StorageBearerTokenChallengeAuthorizationPolicy;
 import com.azure.storage.common.policy.StorageContentValidationDecodingPolicy;
 import com.azure.storage.common.policy.StorageContentValidationEncodingPolicy;
@@ -116,6 +116,8 @@ public final class BuilderHelper {
             policies.add(new AddHeadersPolicy(headers));
         }
         policies.add(new MetadataValidationPolicy());
+        // Normalize ETag condition headers before authentication because Shared Key includes them in the signature.
+        policies.add(new StorageEtagPolicy());
 
         policies.add(new StorageContentValidationEncodingPolicy());
         policies.add(new StorageContentValidationDecodingPolicy());
@@ -145,8 +147,6 @@ public final class BuilderHelper {
         policies.add(getResponseValidationPolicy());
 
         policies.add(new HttpLoggingPolicy(logOptions));
-
-        policies.add(new ScrubEtagPolicy());
 
         return new HttpPipelineBuilder().policies(policies.toArray(new HttpPipelinePolicy[0]))
             .httpClient(httpClient)